+++ thorin/CMakeLists.txt
@@ -36,8 +36,8 @@ add_library(libthorin
     be/emitter.h
     be/dot/dot.cpp
     be/dot/dot.h
-    be/h/bootstrapper.cpp
-    be/h/bootstrapper.h
+    be/h/bootstrap.cpp
+    be/h/bootstrap.h
     fe/ast.cpp
     fe/ast.h
     fe/lexer.cpp
